+++ pages/about/newsletter/index.jsx
@@ -1,6 +1,16 @@
+import {
+  useMemo,
+} from 'react';
+import { useRouter } from 'next/router';
 // components
 import LayoutNewsletter from 'layout/app/newsletter';
 
 export default function NewsletterPage() {
-  return (<LayoutNewsletter />);
+  const {
+    query,
+  } = useRouter();
+
+  const isOceanWatch = useMemo(() => query.origin === 'ocean-watch', [query]);
+
+  return (<LayoutNewsletter isOceanWatch={isOceanWatch} />);
 }